INS-309: move jet coordinator configuration to ledger configuration
